MQTT/publisher/requirements.txt
paho-mqtt
odroid-wiringpi
odroid-dht11
